Task #160337 chore: Moving isClientExist() function  helper to class file
