+++ src/com_tjvendors/admin/models/fields/vendorslist.php
@@ -13,9 +13,10 @@ defined('_JEXEC') or die();
 use Joomla\CMS\Factory;
 use Joomla\CMS\HTML\HTMLHelper;
 use Joomla\CMS\Language\Text;
+use Joomla\CMS\Form\FormHelper;
 
 jimport('joomla.form.helper');
-\FormHelper::loadFieldClass('list');
+FormHelper::loadFieldClass('list');
 
 /**
  * Supports an HTML select list of vendors
