+++ src/com_tjvendors/site/helpers/fronthelper.php
@@ -633,36 +633,4 @@ class TjvendorFrontHelper
 			Factory::getApplication()->enqueueMessage(Text::_('COM_TJVENDORS_DB_EXCEPTION_WARNING_MESSAGE'), 'error');
 		}
 	}
-
-	/**
-	 * Get client name
-	 *
-	 * @param   string   $client    client
-	 * @param   integer  $vendorId  Venodor ID
-	 *
-	 * @return  Boolean Client exist or not
-	 *
-	 * @since  1.3.0
-	 */
-	public function isClientExist($client, $vendorId)
-	{
-		$db = Factory::getDbo();
-		$query = $db->getQuery(true);
-		$query->select('client');
-		$query->from($db->quoteName('#__vendor_client_xref'));
-		$query->where($db->quoteName('client') . ' = ' . $db->quote($client));
-		$query->where($db->quoteName('vendor_id') . ' = ' . (int) $vendorId);
-		$db->setQuery($query);
-
-		$return = $db->loadResult();
-
-		if ($return)
-		{
-			return true;
-		}
-		else
-		{
-			return false;
-		}
-	}
 }
