+++ src/com_tjvendors/site/includes/vendor.php
@@ -615,4 +615,36 @@ class TjvendorsVendor extends CMSObject
 
 		return $total;
 	}
+
+	/**
+	 * Get client name
+	 *
+	 * @param   string   $client    client
+	 * @param   integer  $vendorId  Venodor ID
+	 *
+	 * @return  Boolean Client exist or not
+	 *
+	 * @since  1.3.0
+	 */
+	public function isClientExist($client, $vendorId)
+	{
+		$db = Factory::getDbo();
+		$query = $db->getQuery(true);
+		$query->select('client');
+		$query->from($db->quoteName('#__vendor_client_xref'));
+		$query->where($db->quoteName('client') . ' = ' . $db->quote($client));
+		$query->where($db->quoteName('vendor_id') . ' = ' . (int) $vendorId);
+		$db->setQuery($query);
+
+		$return = $db->loadResult();
+
+		if ($return)
+		{
+			return true;
+		}
+		else
+		{
+			return false;
+		}
+	}
 }
